requirements.txt
torch
torchvision
numpy
scikit-image
opencv-python
pillow
pyyaml
tensorboard
tqdm
lpips